+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ## 7.2.1
+* **improvement** Added tracking for outlinks to MatomoTracker. [#342](https://github.com/matomo-org/matomo-sdk-ios/issues/342)
 * **improvement** Added back detailed phone model feature #253. [#330](https://github.com/matomo-org/matomo-sdk-ios/pull/330)
 * **improvement** Set application extension API only to true #334. [#335](https://github.com/matomo-org/matomo-sdk-ios/pull/335)
 * **bugfix** Fixed an issue where device information and operating system sometimes wasn't tracked. [#329](https://github.com/matomo-org/matomo-sdk-ios/issues/329)